+++ CHANGELOG.rst
@@ -10,8 +10,10 @@ Development Version (unreleased):
 - Fix interoperability issues between netcdf4/h5netcdf, namely resizing variables using partial slices,
 creating string attributes with NULLTERM, proper attachment of _Netcdf4Coordinates and _Netcdf4Dimid as well as special string type fillvalues (:pull:`286`).
   By `Kai Mühlbauer <https://github.com/kmuehlbauer>`_ and `David Huard <https://github.com/huard>`_
-- Add the `format` argument to `h5netcdf.File` and support for the `NETCDF4_CLASSIC` format (:issue:`280`, :pull:``).
+- Add the `format` argument to `h5netcdf.File` and support for the `NETCDF4_CLASSIC` format (:issue:`280`, :pull:`283`).
   By `David Huard <https://github.com/huard>`_
+- Do not return padded arrays for slices larger than variable shape (:issue:`287`, :pull:`288`).
+  By `Kai Mühlbauer <https://github.com/kmuehlbauer>`_
 
 Version 1.6.4 (August 5th, 2025):